python-developer/07-flask/szkolenie-07-jwt-postgresql-testy/requirements.txt
flask
flask-sqlalchemy
PyJWT
pytest
